requirements.txt
faiss-cpu==1.13.2
joblib==1.5.3
kneed==0.8.6
lightfm @ git+https://github.com/lyst/lightfm.git
matplotlib==3.10.8
numpy==2.4.4
overrides==7.7.0
pandas==3.0.2
Requests==2.33.1
scikit-learn==1.8.0
scipy==1.17.1
scikit-surprise==1.1.4
tqdm==4.67.3
